% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{c_zcu}
\alias{c_zcu}
\title{Official university color}
\format{
A character scalar (hex color string).
}
\usage{
c_zcu
}
\description{
The official dark-blue color of the University of West Bohemia as a hex
string. Used as the default single color in plots.
}
\seealso{
\link{c_faks}, \link{c_parts}
}
