% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1.R
\name{ex1}
\alias{ex1}
\title{Bar plot for a single variable}
\usage{
ex1(dat, var)
}
\arguments{
\item{dat}{A name of dataframe.}

\item{var}{A name of variable.}
}
\value{
Horizontal barplot.
}
\description{
(preferably categorical var or numeric var with few values).
}
\examples{
ex1(ggplot2::mpg, manufacturer)
}
