% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex2.R
\name{ex2}
\alias{ex2}
\title{Visual crosstabulation of two vars}
\usage{
ex2(dat, var1, var2)
}
\arguments{
\item{dat}{A name of dataframe.}

\item{var1}{A name of the first variable, which goes into rows}

\item{var2}{A name of the second variable, which goes into columns}
}
\value{
Crosstabulated barplot.
}
\description{
Visual crosstabulation of two vars
}
\examples{
ex2(ggplot2::mpg, manufacturer, cyl)

}
