% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_labs.R
\name{impute_labs}
\alias{impute_labs}
\title{Impute labs from codebook}
\usage{
impute_labs(dat)
}
\arguments{
\item{dat}{A data frame}
}
\value{
A data frame
}
\description{
Internal data prep functions (used in multiple data prep procedures)
}
