% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bat.R
\name{mean_bat}
\alias{mean_bat}
\title{Means of a battery}
\usage{
mean_bat(dat, vars, round_places = 0, na_lab = c("NA", "Bez odpovědďi"))
}
\arguments{
\item{dat}{A data frame}

\item{vars}{A variable}

\item{round_places}{Rounding number}

\item{na_lab}{DNK}
}
\value{
A tibble
}
\description{
Means of a battery
}
