% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bat_gr.R
\name{mean_bat_gr}
\alias{mean_bat_gr}
\title{Means of a battery, grouped}
\usage{
mean_bat_gr(
  dat,
  vars,
  grvar,
  round_places = 0,
  add_total = FALSE,
  lab_total = "ZČU",
  na_lab = c("NA", "Bez odpovědďi")
)
}
\arguments{
\item{dat}{A data frame}

\item{vars}{A variable}

\item{grvar}{A grouping variable}

\item{round_places}{Rounding number}

\item{add_total}{DNK}

\item{lab_total}{DNK}

\item{na_lab}{DNK}
}
\value{
A tibble
}
\description{
Means of a battery, grouped
}
