% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Barplot of a single categorical variable}
\usage{
plot_bar(dat, horiz = TRUE)
}
\arguments{
\item{dat}{A data frame to plot.
It has to include columns named: xvar, yvar, labvar (text labels for yvar), cvar(fill color), title, subtitle, caption.}

\item{horiz}{TRUE for horinzontal plot, FALSE for vertical plot.}
}
\value{
Plot
}
\description{
Barplot of a single categorical variable
}
