% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lolli.R
\name{plot_lolli}
\alias{plot_lolli}
\title{lollipop plot of single categorical variable}
\usage{
plot_lolli(dat, horiz = TRUE)
}
\arguments{
\item{dat}{A data frame.
It has to include columns named: xvar, yvar, labvar (text labels for yvar),
cvar(fill color), title, subtitle, caption.}

\item{horiz}{TRUE for horinzontal plot, FALSE for vertical plot.}
}
\value{
Plot
}
\description{
lollipop plot of single categorical variable
}
