% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stack.R
\name{plot_stack}
\alias{plot_stack}
\title{Plot stacked percentages by groupvar}
\usage{
plot_stack(dat, horiz = TRUE)
}
\arguments{
\item{dat}{A data frame.}

\item{horiz}{TRUE for horinzontal plot, FALSE for vertical plot.}
}
\value{
Plot.
}
\description{
Plot stacked percentages by groupvar
}
