% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_var.R
\name{polish_var}
\alias{polish_var}
\title{Generate final version xvar or zvar to be used in plots}
\usage{
polish_var(
  dat,
  var_origin = "xvar_df",
  var_final = "xvar",
  wrap = TRUE,
  chrnum = .uwb_vals$chrnum,
  nsize = TRUE
)
}
\arguments{
\item{dat}{A data frame}

\item{var_origin}{DNK}

\item{var_final}{DNK}

\item{wrap}{DNK}

\item{chrnum}{DNK}

\item{nsize}{DNK}
}
\value{
DNK
}
\description{
It allows wrapping long text and addding nsize info
result_name:
xvar_df/zvar = default, basic text version as found in the data set
xzvar_nr = helper var with level number, to be used for reordering the modified version, removed
xvar/zvar = final version that goes to the plot (wrapped or not, with or without nsize)
}
