% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bat.R
\name{prep_bat}
\alias{prep_bat}
\title{Batteries data prep}
\usage{
prep_bat(
  dat,
  vars = names(dat),
  drop_na = TRUE,
  x_wrap = TRUE,
  x_chrnum = .uwb_vals$chrnum,
  x_nsize = FALSE
)
}
\arguments{
\item{dat}{A data frame in a wide format}

\item{vars}{A variable}

\item{drop_na}{Drop NA or not?}

\item{x_wrap}{DNK}

\item{x_chrnum}{DNK}

\item{x_nsize}{DNK}
}
\value{
A tibble
}
\description{
Dat is wide data, it is pivoted longer inside the function
}
