% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gr.R
\name{prep_gr}
\alias{prep_gr}
\title{Grouped data prep with or without total}
\usage{
prep_gr(
  dat,
  var,
  grvar,
  drop_na = TRUE,
  add_total = FALSE,
  lab_total = "ZČU",
  grvar_to_x = TRUE,
  show_nsize = TRUE,
  x_wrap = TRUE,
  x_chrnum = .uwb_vals$chrnum,
  z_wrap = TRUE,
  z_chrnum = .uwb_vals$chrnum
)
}
\arguments{
\item{dat}{A data frame}

\item{var}{A variable}

\item{grvar}{A group variable}

\item{drop_na}{Drop NA or not?}

\item{add_total}{T/F}

\item{lab_total}{DNK}

\item{grvar_to_x}{DNK}

\item{show_nsize}{DNK}

\item{x_wrap}{DNK}

\item{x_chrnum}{DNK}

\item{z_wrap}{DNK}

\item{z_chrnum}{DNK}
}
\value{
A data frame
}
\description{
var becomes zvar, grvar becomes xvar, it can be switched by grvar_to_x = FALSE
}
\examples{
prep_gr(ggplot2::mpg, class, cyl)

}
