% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_mc.R
\name{prep_mc}
\alias{prep_mc}
\title{Multiple choice data prep}
\usage{
prep_mc(
  dat,
  vars = names(dat),
  chosen,
  drop_na = TRUE,
  order = FALSE,
  x_wrap = TRUE,
  x_chrnum = .uwb_vals$chrnum,
  x_nsize = TRUE
)
}
\arguments{
\item{dat}{A data frame}

\item{vars}{Variables}

\item{chosen}{DNK}

\item{drop_na}{Drop NA or not?}

\item{order}{Order results or not?}

\item{x_wrap}{DNK}

\item{x_chrnum}{DNK}

\item{x_nsize}{DNK}
}
\value{
A tibble
}
\description{
(= batttery with 2 options: chosen vs. not chosen)
}
\examples{
prep_mc(ggplot2::mpg |> dplyr::select(manufacturer), chosen = "audi")

}
