% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_mc_gr.R
\name{prep_mc_gr}
\alias{prep_mc_gr}
\title{Grouped multiple choice data prep}
\usage{
prep_mc_gr(
  dat,
  vars = names(dat),
  grvar,
  chosen,
  drop_na = TRUE,
  add_total = FALSE,
  lab_total = "ZČU",
  x_wrap = TRUE,
  x_chrnum = .uwb_vals$chrnum,
  x_nsize = TRUE
)
}
\arguments{
\item{dat}{A data frame}

\item{vars}{A variable}

\item{grvar}{A grouping variable}

\item{chosen}{A value}

\item{drop_na}{Drop NA or not?}

\item{add_total}{Add total or not?}

\item{lab_total}{DNK}

\item{x_wrap}{DNK}

\item{x_chrnum}{DNK}

\item{x_nsize}{DNK}
}
\value{
A tibble
}
\description{
grvar becomes xvar
}
\examples{
"DNK"

}
