% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_mc_long_faks.R
\name{prep_mc_long_faks}
\alias{prep_mc_long_faks}
\title{Prep MC data in long format}
\usage{
prep_mc_long_faks(dat)
}
\arguments{
\item{dat}{A data frame}
}
\value{
A tibble
}
\description{
total + grouped by faculty
}
