% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_single.R
\name{prep_single}
\alias{prep_single}
\title{Prepare percentage data of single var}
\usage{
prep_single(dat, var, order = FALSE, drop_na = TRUE)
}
\arguments{
\item{dat}{A data frame}

\item{var}{A variable}

\item{order}{Should the results be ordered?}

\item{drop_na}{Drop NA or not?}
}
\value{
A tibble with enhanced frequency table
}
\description{
Prepare percentage data of single var
}
\examples{
prep_single(ggplot2::mpg, cyl)

}
