% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_uwb.R
\name{scale_color_uwb}
\alias{scale_color_uwb}
\title{scale_color_uwb}
\usage{
scale_color_uwb(name, n, add_na = FALSE)
}
\arguments{
\item{name}{A name of the palette}

\item{n}{Number of colors}

\item{add_na}{Add color for unknown category?}
}
\description{
scale_color_uwb
}
