% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwb_palettes.R
\name{uwb_palettes}
\alias{uwb_palettes}
\title{Create a uwb pallete}
\usage{
uwb_palettes(
  name = "quali",
  n,
  all_palettes = .uwb_scales,
  type = c("discrete", "continuous"),
  add_na = FALSE
)
}
\arguments{
\item{name}{A name of the palette}

\item{n}{Number of colors}

\item{all_palettes}{DNK}

\item{type}{discrete or continuous}

\item{add_na}{Add color for unknown category?}
}
\description{
create a palette from the provided colors (from the list uwb_scales)
}
\examples{
uwb_palettes(name = "quali", n = 3, type = "discrete")
uwb_palettes(n = 5, add_na = TRUE)

}
