% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_textlabs.R
\name{generate_textlabs}
\alias{generate_textlabs}
\title{Generate vars to be used for geom_text}
\usage{
generate_textlabs(dat, round_places = 0)
}
\arguments{
\item{dat}{A data frame}

\item{round_places}{Number of decimals}
}
\value{
A data frame
}
\description{
Generate vars to be used for geom_text
}
