% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bat_gr2.R
\name{mean_bat_gr2}
\alias{mean_bat_gr2}
\title{Means of a battery, grouped by two variables}
\usage{
mean_bat_gr2(
  dat,
  vars,
  grvar,
  grvar2,
  round_places = 0,
  na_lab = c("NA", "Bez odpovědďi")
)
}
\arguments{
\item{dat}{A data frame}

\item{vars}{A variable}

\item{grvar}{The first grouping variable}

\item{grvar2}{The second grouping variable}

\item{round_places}{Rounding number}

\item{na_lab}{DNK}
}
\value{
A tibble
}
\description{
Means of a battery, grouped by two variables
}
