% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dodge.R
\name{plot_dodge}
\alias{plot_dodge}
\title{Plot dodge percentages by groupvar}
\usage{
plot_dodge(dat, horiz = TRUE)
}
\arguments{
\item{dat}{A data frame.}

\item{horiz}{TRUE for horinzontal plot, FALSE for vertical plot.}
}
\value{
Plot.
}
\description{
Plot dodge percentages by groupvar
}
