% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_out.R
\name{plot_out}
\alias{plot_out}
\title{Plot points: total value + groups with highlighted outliers (jittered)
Default color scheme is faks because this is the most common use case
Data has to be grouped in order to compute differences from the main trend, the default grouping is by xvar
Data has to include groups + total}
\usage{
plot_out(
  dat,
  horiz = FALSE,
  total = "ZČU",
  c_total = .uwb_scales$faks[10],
  add_line = TRUE,
  out,
  jitter_w = 0.3,
  alpha_shadow = 0.15,
  seed = 123456
)
}
\arguments{
\item{dat}{A data frame.}

\item{horiz}{TRUE for horinzontal plot, FALSE for vertical plot.}

\item{total}{DNK}

\item{c_total}{DNK}

\item{add_line}{TRUE is for trends, it adds a trend line}

\item{out}{DNK}

\item{jitter_w}{DNK}

\item{alpha_shadow}{DNK}

\item{seed}{Random seed number.}
}
\value{
Plot
}
\description{
Plot points: total value + groups with highlighted outliers (jittered)
Default color scheme is faks because this is the most common use case
Data has to be grouped in order to compute differences from the main trend, the default grouping is by xvar
Data has to include groups + total
}
