% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ribbon.R
\name{plot_ribbon}
\alias{plot_ribbon}
\title{Function to generate ribbon plot}
\usage{
plot_ribbon(dat, space = 5, fill = .uwb_scales$quali, margin = 15)
}
\arguments{
\item{dat}{A data frame.}

\item{space}{DNK}

\item{fill}{DNK}

\item{margin}{DNK}
}
\value{
Plot.
}
\description{
hide_lab = do not show labels of small bumps
}
