% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_gr.R
\name{plot_trend_gr}
\alias{plot_trend_gr}
\title{Plot trend by a grouping var}
\usage{
plot_trend_gr(dat, total, add_points = TRUE)
}
\arguments{
\item{dat}{A data frame.}

\item{total}{DNK}

\item{add_points}{DNL}
}
\value{
Plot.
}
\description{
xvar is the time var, zvar is the groups var
}
