% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bat_gr.R
\name{prep_bat_gr}
\alias{prep_bat_gr}
\title{Grouped battery data prep}
\usage{
prep_bat_gr(
  dat,
  vars = names(dat),
  grvar,
  add_total = FALSE,
  lab_total = "ZČU"
)
}
\arguments{
\item{dat}{A data frame}

\item{vars}{A variable}

\item{grvar}{A grouping variable}

\item{add_total}{DNK}

\item{lab_total}{DNK}
}
\value{
A tibble
}
\description{
zvar is value for battery items
}
