% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gr2.R
\name{prep_gr2}
\alias{prep_gr2}
\title{Grouped data prep - two levels of grouping}
\usage{
prep_gr2(
  dat,
  var,
  grvar,
  grvar2,
  drop_na = TRUE,
  add_total = FALSE,
  lab_total = "ZČU",
  grvar_to_x = TRUE,
  show_nsize = TRUE,
  x_wrap = TRUE,
  x_chrnum = .uwb_vals$chrnum,
  z_wrap = TRUE,
  z_chrnum = .uwb_vals$chrnum
)
}
\arguments{
\item{dat}{A data frame}

\item{var}{A variable}

\item{grvar}{The first grouped variable}

\item{grvar2}{The second grouped variable}

\item{drop_na}{Drop NA or not?}

\item{add_total}{Add total or not?}

\item{lab_total}{DNK}

\item{grvar_to_x}{DNK}

\item{show_nsize}{DNK}

\item{x_wrap}{DNK}

\item{x_chrnum}{DNK}

\item{z_wrap}{DNK}

\item{z_chrnum}{DNK}
}
\value{
A tibble
}
\description{
grvar becomes xvar? grvar2 is macro grouping var for facets
}
\examples{
prep_gr2(ggplot2::mpg, cyl, manufacturer, fl)

}
