% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_uwb.R
\name{scale_fill_uwb}
\alias{scale_fill_uwb}
\title{scale_fill_uwb}
\usage{
scale_fill_uwb(name = "quali", n, add_na = FALSE)
}
\arguments{
\item{name}{A name of the palette}

\item{n}{Number of colors}

\item{add_na}{Add color for unknown category?}
}
\description{
scale_fill_uwb
}
